fix: refactor validation, make helper private and remove interface for validation
